fix/QA-116: 9,999,999,999원 입력 시 오류 발생

+++ feature/sent/src/main/java/com/susu/feature/envelope/component/EnvelopeHistoryItem.kt
@@ -76,7 +76,7 @@ fun EnvelopeHistoryItem(
         )
         Spacer(modifier = modifier.weight(1f))
         Text(
-            text = money.toInt().toMoneyFormat() + stringResource(R.string.sent_envelope_card_money_won),
+            text = money.toMoneyFormat() + stringResource(R.string.sent_envelope_card_money_won),
             style = SusuTheme.typography.title_xs,
             color = if (type == EnvelopeType.SENT.name) Gray100 else Gray50,
         )
